+++ backend/src/requestTypes.ts
@@ -1,5 +1,5 @@
 import { Request, Response, NextFunction } from "express";
-import { User } from "../models/user";
+import { User } from "./user/user.type";
 
 export interface RequestHandler<P = {}, B = {}, Q = {}> {
   (req: Request<P, any, B, Q>, res: Response, next?: NextFunction):
