+++ backend/src/user/user.route.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { userController } from "../controllers/userController";
+import { userController } from "./user.controller";
 
 const router = Router();
